Modify prabhav-bot
