--- bots/prabhav-bot-2/run.sh
@@ -1,2 +1,0 @@
-cd ../bots/prabhav-bot-2/
-python3 MyBot.py
